shows more details in detailview and can go to maps

+++ app/src/main/java/com/example/favouriteplaces/Place.kt
@@ -5,10 +5,10 @@ import com.google.firebase.firestore.DocumentId
 data class Place( @DocumentId var docID: String? = null,
     val title: String? = null,
     var description: String? = null,
-    val lat: Long? = null,
-    val lng: Long? = null,
+    val lat: Double? = null,
+    val lng: Double? = null,
     val category: String? = null,
-    val stars: Double? = null,
+    val stars: Float? = null,
     val review: String? = null,
     val public: Boolean? = null,
     val author: String? = null
